requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==8.13.2
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
migrate==0.3.7
packaging==21.0
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
